Refactored cell renderers to more closely mimic cell editor structure
